embedding-service/requirements.txt
# Batman AI - Embedding Service (Docker/GPU)
# Pinned to versions compatible with PyTorch 2.1.0 (cuda11.8)
# transformers 4.57+ breaks on PyTorch 2.1.0 due to pytree API change

fastapi==0.115.6
uvicorn[standard]==0.34.0
sentence-transformers==2.7.0
transformers==4.44.2
faiss-cpu==1.9.0.post1
numpy==1.26.4
pydantic==2.10.4
